* Upload of version 0.8.0 beta1
git-svn-id: https://plugins.svn.wordpress.org/sns-count-cache/trunk@1225141 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ includes/class-follow-second-cache-engine.php
@@ -201,7 +201,7 @@ class Follow_Second_Cache_Engine extends Cache_Engine {
 	  	
 		foreach ( $this->target_sns as $sns => $active ) {					  					  
 			if ( $active ) {
-			  	$sns_followers[$sns] = -1;
+			  	$sns_followers[$sns] = (int) -1;
 			}
 		}
